requirements.txt
annotated-types==0.7.0
anyascii==0.3.2
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
beautifulsoup4==4.11.2
certifi==2023.7.22
charset-normalizer==3.2.0
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.5
django-bulk-update==2.2.0
django-cors-headers==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.3
django-modelcluster==6.0
django-ninja==1.3.0
django-permissionedforms==0.1
django-taggit==4.0.0
django-treebeard==4.7
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
filetype==1.2.0
gunicorn==21.2.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
ipython==8.16.0
jedi==0.19.0
jupyter_client==8.3.1
jupyter_core==5.3.2
l18n==2021.3
matplotlib-inline==0.1.6
mysql-connector-python==8.1.0
mysqlclient==2.2.0
nest-asyncio==1.5.8
numpy==1.26.0
openpyxl==3.1.2
orjson==3.10.7
packaging==23.1
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pillow-heif==0.13.0
pkgconfig==1.5.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.16.1
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
telepath==0.3.1
tornado==6.3.3
traitlets==5.10.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.0
wagtail==5.1.1
wcwidth==0.2.7
webencodings==0.5.1
whitenoise==6.5.0
Willow==1.6.2
zipp==3.17.0
